try to force tar format and uid/gid
If the uid and gid of the files on disk is too large, tar will silently
upgrade the format of the tar file to include pax headers.  This will
lead to warnings or errors on decompressing.  Try to force the format to
prevent this.  This may cause the archive creation to fail, but there is
an extra check to validate the tarball later.  If using gnu tar, we can
force the uid and gid, which should prevent this issue.