move Makefile.PL aside for now
The default way of using Distar will put it in a subdirectory of the
dist.  If the dist doesn't specify NO_RECURS, EUMM will try to include
Makefile.PL files in subdirectories.  This means it will include Distar
if it has one.

For now, move the Makefile.PL aside.  This will be revisited in the
future.